Add models and routers for SearchWords module
