+++ Source/Models/User.swift
@@ -1,0 +1,13 @@
+//
+//  User.swift
+//  Find-and-Learn
+//
+//  Created by Руслан on 21.03.2022.
+//
+
+import Foundation
+
+struct User {
+    let email: String
+    let password: String
+}
