--- Source/Scenarios/Entrance/Login/LoginViewController.swift
@@ -1,14 +1,0 @@
-//
-//  LoginViewController.swift
-//  Find-and-Learn
-//
-//  Created by Руслан on 19.03.2022.
-//
-
-import UIKit
-
-class LoginViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
