--- Source/Scenarios/Entrance/Registration/RegistrationViewController.swift
@@ -1,15 +1,0 @@
-//
-//  RegistrationViewController.swift
-//  Find-and-Learn
-//
-//  Created by Руслан on 24.02.2022.
-//
-
-import UIKit
-
-final class RegistrationViewController: UIViewController {
-    // MARK: - Lifecycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
